Refactor utils package structure and update imports
Moved several utility classes to more specific subpackages (dto, format, function) for better organization. Updated all affected imports throughout the codebase to reflect the new package structure.
