+++ src/main/java/nl/radiantrealm/library/utils/function/ThrowingBiFunction.java
@@ -1,4 +1,4 @@
-package nl.radiantrealm.library.utils;
+package nl.radiantrealm.library.utils.function;
 
 public interface ThrowingBiFunction<T, U, R> {
     R apply(T t, U u) throws Exception;
