+++ src/main/java/nl/radiantrealm/library/utils/function/ThrowingFunction.java
@@ -1,4 +1,4 @@
-package nl.radiantrealm.library.utils;
+package nl.radiantrealm.library.utils.function;
 
 public interface ThrowingFunction<T, R> {
     R apply(T t) throws Exception;
